requirements.txt
arcticdb
pandas
ccxt
fredapi
yfinance
python-dotenv
